// src/app/core/domain/models/step1.model.ts
export interface SubArray1Item {
    field1: string;
    field2: string;
    field3: string;
    field4: string;
    field5: string;
}

export interface SubArray2Item {
    field1: string;
    field2: string;
    field3: string;
    field4: string;
    field5: string;
}


export interface Step1Model {
    field1: string;
    field2: string;
    subArray1: SubArray1Item[];
    subArray2: SubArray2Item[];
}